Move index.md contents to README
I've removed index.md, and moved all of its content to the README. This has a few benefits:

1. anyone coming in through GitHub will be able to see the main point of this repository - resources - immediately, without a confusing click on 'index.md'.
2. this automatically deploys to /resources/, without the need for a redirect or a template variable
3. This conforms to other README lists which exist on GitHub, which allow users to interface with it more naturally.
4. This fixes a broken build (see https://github.com/opensourcedesign/opensourcedesign.github.io/issues/67).

The detractions, that I can think of, are that the README no longer explains what this repository is as it's _only_ function. However, I don't think that needs to be the only function of a README - and, I think, if the repository can self-describe by stating what it is by showing you what it is, people can intuit the focus of the repository more easily.

This is related to and stems from https://github.com/opensourcedesign/opensourcedesign.github.io/issues/67.